PassiveIP check configurable; throws an exception for insecure connects
This fix provides the configuration of the passive IP check. merged from
remote 1.0.x.
Additionally the connects of insecure IPs on a data connection are
reported by an FTPException (in contrast to returning bull in the
original implementation) to avoid empty files being created by opening
of internal output streams.