Provides enhanced log message for passive client IP check
The check for the client IP address connecting to the passive listener
being the same as the one having sent the PASV command can be activated
and deactivated. 
Users running with deactivated the check (temporarily) need a log
message in this case, too, as they might want to identify the affected
clients by this log message.