More strict EOL conversion in ASCII mode
In case a CRLF file is uploaded from a UNIX FTP Client to the Apache FTP
Server the output file is corrupted. The CRLF is converted to CRLFCRLF,
which is wrong. See details in RFC 959, chapter 3.1.1.1, p.11
https://tools.ietf.org/html/rfc959#page-11

The strict mode must be activated by configuration, the old behavior is
kept for the sake of backwards compatibility.